moved listener to separate package
