+++ .gitignore
@@ -1,12 +1,11 @@
+target/
+*.class
 
+# Package Files #
+*.jar
+*.war
+*.ear
+
+#IDE
+*.iml
 .idea/
-/target/*
-/test-output/*
-*.log
-*.orig
-.DS_Store
-.checkstyle
-.classpath
-.idea
-.project
-.settings
