+++ src/main/java/listener/ScreenshotListener.java
@@ -1,5 +1,6 @@
-package appLogic;
+package listener;
 
+import appLogic.ApplicationManager;
 import com.google.common.io.Files;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
